+++ accessories/motion-accessory.js
@@ -32,6 +32,7 @@ function CBusMotionAccessory(platform, accessoryData) {
 
 CBusMotionAccessory.prototype.getMotionState = function (callback) {
 	this.client.receiveLevel(this.netId, message => {
+		if (this._handleClientResponseError(message, callback, 'getMotionState')) return;
 		this._log(FILE_ID, `getState`, message.level);
 		callback(false, /* state: */ message.level ? 1 : 0);
 	});